+++ apina-cli/src/main/kotlin/fi/evident/apina/cli/Apina.kt
@@ -30,6 +30,7 @@ object Apina {
             processor.settings.platform = arguments.platform
             processor.settings.typeWriteMode = arguments.typeWriteMode
             processor.settings.optionalTypeMode = arguments.optionalTypeMode
+            processor.settings.reexportImports = arguments.reexportImports
 
             for (blackBoxPattern in arguments.blackBoxPatterns)
                 processor.settings.blackBoxClasses.addPattern(blackBoxPattern)