+++ apina-core/src/main/kotlin/fi/evident/apina/output/ts/AbstractTypeScriptGenerator.kt
@@ -61,6 +61,13 @@ abstract class AbstractTypeScriptGenerator(
 
             out.writeLine()
         }
+
+        if (settings.reexportImports && imports.isNotEmpty()) {
+            for (anImport in imports)
+                out.writeExport(anImport.types.map { it.name })
+
+            out.writeLine()
+        }
     }
 
     private fun writeTypes() {